+++ Magic/src/main/java/com/elmakers/mine/bukkit/protection/WorldGuardFlagsManager.java
@@ -19,10 +19,13 @@ import com.sk89q.worldguard.protection.flags.StringFlag;
 import com.sk89q.worldguard.protection.flags.registry.FlagRegistry;
 
 public class WorldGuardFlagsManager implements WorldGuardFlags {
+    public static SetFlag<String> ALWAYS_ALLOWED_SPELLS = new SetFlag<>("always-allowed-spells", RegionGroup.ALL, new StringFlag(null));
     public static SetFlag<String> ALLOWED_SPELLS = new SetFlag<>("allowed-spells", RegionGroup.ALL, new StringFlag(null));
     public static SetFlag<String> BLOCKED_SPELLS = new SetFlag<>("blocked-spells", RegionGroup.ALL, new StringFlag(null));
+    public static SetFlag<String> ALWAYS_ALLOWED_SPELL_CATEGORIES = new SetFlag<>("always-allowed-spell-categories", RegionGroup.ALL, new StringFlag(null));
     public static SetFlag<String> ALLOWED_SPELL_CATEGORIES = new SetFlag<>("allowed-spell-categories", RegionGroup.ALL, new StringFlag(null));
     public static SetFlag<String> BLOCKED_SPELL_CATEGORIES = new SetFlag<>("blocked-spell-categories", RegionGroup.ALL, new StringFlag(null));
+    public static SetFlag<String> ALWAYS_ALLOWED_WANDS = new SetFlag<>("always-allowed-wands", RegionGroup.ALL, new StringFlag(null));
     public static SetFlag<String> ALLOWED_WANDS = new SetFlag<>("allowed-wands", RegionGroup.ALL, new StringFlag(null));
     public static SetFlag<String> BLOCKED_WANDS = new SetFlag<>("blocked-wands", RegionGroup.ALL, new StringFlag(null));
     public static SetFlag<String> SPELL_OVERRIDES = new SetFlag<>("spell-overrides", RegionGroup.ALL, new StringFlag(null));
@@ -46,16 +49,19 @@ public class WorldGuardFlagsManager implements WorldGuardFlags {
             callingPlugin.getLogger().warning("Failed to find FlagRegistry, custom WorldGuard flags will not work");
         }
         registry.register(ALLOWED_SPELLS);
+        registry.register(ALWAYS_ALLOWED_SPELLS);
         registry.register(BLOCKED_SPELLS);
+        registry.register(ALWAYS_ALLOWED_SPELL_CATEGORIES);
         registry.register(ALLOWED_SPELL_CATEGORIES);
         registry.register(BLOCKED_SPELL_CATEGORIES);
+        registry.register(ALWAYS_ALLOWED_WANDS);
         registry.register(ALLOWED_WANDS);
         registry.register(BLOCKED_WANDS);
         registry.register(SPELL_OVERRIDES);
         registry.register(DESTRUCTIBLE);
         registry.register(REFLECTIVE);
         registry.register(SPAWN_TAGS);
-        callingPlugin.getLogger().info("Registered custom WorldGuard flags: allowed-spells, blocked-spells, allowed-spell-categories, blocked-spell-categories, allowed-wands, blocked-wands, spell-overrides, destructible, reflective, spawn-tags");
+        callingPlugin.getLogger().info("Registered custom WorldGuard flags: allowed-spells, always-allowed-spells, blocked-spells, always-allowed-spell-categories, allowed-spell-categories, blocked-spell-categories, allowed-wands, always-allowed-wands, blocked-wands, spell-overrides, destructible, reflective, spawn-tags");
     }
 
     @Nullable
@@ -80,6 +86,10 @@ public class WorldGuardFlagsManager implements WorldGuardFlags {
     @Override
     public Boolean getWandPermission(RegionAssociable source, ApplicableRegionSet checkSet, Wand wand) {
         String wandTemplate = wand.getTemplateKey();
+
+        Set<String> alwaysAllowed = checkSet.queryValue(source, ALWAYS_ALLOWED_WANDS);
+        if (alwaysAllowed != null && (alwaysAllowed.contains("*") || alwaysAllowed.contains(wandTemplate))) return true;
+
         Set<String> blocked = checkSet.queryValue(source, BLOCKED_WANDS);
         if (blocked != null && blocked.contains(wandTemplate)) return false;
 
@@ -95,6 +105,12 @@ public class WorldGuardFlagsManager implements WorldGuardFlags {
     @Override
     public Boolean getCastPermission(RegionAssociable source, ApplicableRegionSet checkSet, SpellTemplate spell) {
         String spellKey = spell.getSpellKey().getBaseKey();
+
+        Set<String> alwaysAllowed = checkSet.queryValue(source, ALWAYS_ALLOWED_SPELLS);
+        if (alwaysAllowed != null && (alwaysAllowed.contains("*") || alwaysAllowed.contains(spellKey))) return true;
+
+        Set<String> alwaysAllowedCategories = checkSet.queryValue(source, ALWAYS_ALLOWED_SPELL_CATEGORIES);
+        if (alwaysAllowedCategories != null && spell.hasAnyTag(alwaysAllowedCategories)) return true;
 
         Set<String> blocked = checkSet.queryValue(source, BLOCKED_SPELLS);
         if (blocked != null && blocked.contains(spellKey)) return false;